Add files via upload
It now reads from both types of sensors (DHT22 and DS18B20) and reads all six sensors attached to the pi. The DHT22 was giving lots of spikes in the data so module to check for spikes was added. Dewpoint can also be calculated and sent to DataDog but a new metric for Dewpoint will have to be created first. Feel free to fix any bad coding, I'm still learning.